Use scipy's wasserstein_distance to accelerate
Calculating the Wasserstein distance using the optimization method is too slow, so here I replaced the optimization method with the direct calculation method implemented in Scipy. The improvement is significant. The optimization method did not complete the 4-D data clustering task of about 20k in half an hour, while the direct approach did it in 5 seconds.